implements author able to modify own notes only

+++ pages/api/notes/[noteId]/index.ts
@@ -1,7 +1,9 @@
 import { NextApiHandler, NextApiRequest, NextApiResponse } from "next";
+import { Session } from "next-auth";
 import { getSession } from "next-auth/react";
 import db from "../../../../config/dbConnect";
 import Note from "../../../../models/Note";
+import { checkAuthor } from "../../../../utils/checkAuthor";
 
 const getNote = async (req: NextApiRequest, res: NextApiResponse) => {
   try {
@@ -18,27 +20,46 @@ const getNote = async (req: NextApiRequest, res: NextApiResponse) => {
   }
 };
 
-const updateNote = async (req: NextApiRequest, res: NextApiResponse) => {
+const updateNote = async (
+  req: NextApiRequest,
+  res: NextApiResponse,
+  session: Session
+) => {
   try {
     await db.connect();
-    const note = await Note.findByIdAndUpdate(req.query.noteId, req.body, {
-      new: true,
-    });
-    await db.disconnect();
+    let note = await Note.findById(req.query.noteId);
+
     if (!note) {
       res.status(404).json({ error: "Note not found." });
-    } else {
-      res.status(200).json({ note });
     }
+
+    checkAuthor(res, note, session);
+
+    note = await Note.findByIdAndUpdate(req.query.noteId, req.body, {
+      new: true,
+    });
+
+    await db.disconnect();
+
+    res.status(200).json({ note });
   } catch (error) {
     res.status(500).json({ error });
   }
 };
 
-const deleteNote = async (req: NextApiRequest, res: NextApiResponse) => {
+const deleteNote = async (
+  req: NextApiRequest,
+  res: NextApiResponse,
+  session: Session
+) => {
   try {
     await db.connect();
-    const note = await Note.findByIdAndDelete(req.query.noteId);
+    let note = await Note.findById(req.query.noteId);
+    if (!note) {
+      return res.status(404).json({ error: "Note not found" });
+    }
+    checkAuthor(res, note, session);
+    await Note.findByIdAndDelete(req.query.noteId);
     await db.disconnect();
     res.status(200).json({ message: "Note deleted successfully" });
   } catch (error) {
@@ -63,10 +84,10 @@ const handler: NextApiHandler = async (
       await getNote(req, res);
       break;
     case "PATCH":
-      await updateNote(req, res);
+      await updateNote(req, res, session);
       break;
     case "DELETE":
-      await deleteNote(req, res);
+      await deleteNote(req, res, session);
       break;
     default:
       res
